Ultimas modificações

+++ script.js
@@ -24,7 +24,11 @@ formList.addEventListener('submit', (event) => {
 
     //Recuperando valor del input
     const _assignment = document.querySelector("#tarefa").value
+
+    // alerta 
     const _alert = document.querySelector(".alert")
+
+    // Validação da primera letra do input para criação do id fututamente
     switch (_assignment[0]) {
         case undefined:
         case '1':
@@ -37,32 +41,30 @@ formList.addEventListener('submit', (event) => {
         case '8':
         case '9':
         case '0':
-
+            // Alerta caso o valor esteja vazio ou começe com algum numero
             _alert.style = "background: #d12438; display: block;"
             _alert.innerHTML = `<i class="fas fa-exclamation-circle"></i> Digite um nome valido`
             break;
 
         default:
-
+            // Criando uma tarefa e mostrando a lista de tarefas
             if (event.submitter.id == 'add') {
                 createItem(_assignment)
                 ShowItemList(true)
+
+                // Resetando o formulario
                 formList.reset()
 
-
+                // Alerta de Tarefa Adicionada com sucesso
                 _alert.style = "background: #4faa0e; display: block;"
                 _alert.innerHTML = `<i class="fas fa-check-circle"></i> Adicionada com sucesso`
 
 
             }
-
-
             break;
     }
 
-    //  verificar se o valor nao esta vazio e nao começa com numero
-
-
+    // fecha o formulario de cadastro de tarefa
     if (event.submitter.id == 'hidden') {
         cardForm.style = "display: none; "
         _alert.style = "display: none;"
@@ -76,14 +78,16 @@ formList.addEventListener('submit', (event) => {
 // Mosutrando Item na lista de conteudo
 function ShowItemList(clearList = false) {
 
+    // limpando lista
     if (clearList) {
         _showItemContent.innerHTML = ''
     }
 
-    // Adiciona os item na lista
-    _listToDo.forEach(item => {
-        if (item.state == false) {
-            _showItemContent.innerHTML += `<div  class="cardItem"  >
+    if (_listToDo.length > 0) {
+        // Adiciona os item na lista
+        _listToDo.forEach(item => {
+            if (item.state == false) {
+                _showItemContent.innerHTML += `<div  class="cardItem"  >
                  <input type="checkbox" name="itemList" id="${item.name}" onclick="changeText('${item.name}','t${item.name}')"/> 
                  <label for="${item.name}"  id="t${item.name}"> ${item.name} </label>
                  <button onClick="deleteItem('${item.name}')">
@@ -91,8 +95,8 @@ function ShowItemList(clearList = false) {
                  </button>
 
                  </div>`
-        } else {
-            _showItemContent.innerHTML += `<div  class="cardItem"  >
+            } else {
+                _showItemContent.innerHTML += `<div  class="cardItem"  >
                  <input type="checkbox" name="itemList" id="${item.name}" checked onclick="changeText('${item.name}','t${item.name}')"/> 
                  <label for="${item.name}"  id="t${item.name}" style="text-decoration: line-through;'"> ${item.name} </label>
                  <button onClick="deleteItem('${item.name}')">
@@ -101,8 +105,12 @@ function ShowItemList(clearList = false) {
 
                  </div>`
 
-        }
-    });
+            }
+        });
+    } else {
+        _showItemContent.innerHTML = ' Sem tarefas Diponiveis'
+
+    }
 
 }
 
@@ -110,12 +118,8 @@ function ShowItemList(clearList = false) {
 function deleteItem(id) {
 
     for (let i = 0; i < _listToDo.length; i++) {
-
-
         if (_listToDo[i].name == id) {
-
             const index = _listToDo.indexOf(_listToDo[i])
-
             _listToDo.splice(index, 1)
             ShowItemList(true)
             saveList()
@@ -127,15 +131,14 @@ function deleteItem(id) {
 // salva a lista no localStorage
 function saveList() {
     localStorage.setItem('List_Todo', JSON.stringify(_listToDo))
-
 }
 
-
+// Mostra o formulario de cadastro das tarefas
 function createTask() {
-
     cardForm.style = "display: flex; "
 }
 
+// Muda a formatação do nome da tarefa
 function changeText(id, idText) {
     const _state = document.querySelector(`#${id}`)
     const _textItem = document.querySelector(`#${idText}`)
@@ -158,5 +161,6 @@ function changeText(id, idText) {
 
     }
 }
+
 
 window.onload = ShowItemList()
